% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_sens_h2_nonnorm_exposure.R
\name{mr_sens_h2_nonnorm_exposure}
\alias{mr_sens_h2_nonnorm_exposure}
\title{Sensitivity analysis}
\usage{
mr_sens_h2_nonnorm_exposure(
  harmonized_data,
  K = 5,
  method_list = c("mr_ivw", "mr_simple_mode", "mr_weighted_mode", "mr_weighted_median",
    "mr_egger_regression")
)
}
\arguments{
\item{harmonized_data}{output of TwoSampleMR::harmonise_data()}

\item{K}{number of instrument groups, default = 5}

\item{method_list}{MR methods, selected from TwoSampleMR::mr_method_list()}
}
\value{
list containing \itemize{
\item \code{result} results of MR sensitivity analysis
\item \code{instrument} summary of instruments and group membership
\item \code{Q} Cochran’s Q statistic
\item \code{I2} Higgins & Thompson’s I2 statistic
\item \code{het_P} heterogeneity p-value
}
}
\description{
Rank core instruments by significance (per-variant heritability, exposure is not normalized to have a unit variance)
}
