% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRSensiCore-package.R
\docType{package}
\name{MRSensiCore-package}
\alias{MRSensiCore}
\alias{MRSensiCore-package}
\title{MRSensiCore: Mendelian randomization sensitivity analysis based on omnigenic model and core gene assumption}
\description{
Mendelian randomization sensitivity analysis
}
\author{
\strong{Maintainer}: Tianyuan Lu \email{tianyuan.lu@mail.mcgill.ca} (\href{https://orcid.org/0000-0002-5664-5698}{ORCID})

}
\keyword{internal}
