% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_core_instrument.R
\name{permute_core_instrument}
\alias{permute_core_instrument}
\title{Permutation test}
\usage{
permute_core_instrument(
  harmonized_data,
  mr_sens_result,
  core_group = 1,
  exclude_group = 1,
  Nperm = 1000,
  method_list = c("mr_ivw", "mr_weighted_mode", "mr_weighted_median",
    "mr_penalised_weighted_median", "mr_egger_regression")
)
}
\arguments{
\item{harmonized_data}{output of TwoSampleMR::harmonise_data()}

\item{mr_sens_result}{output of \link{mr_sens_h2}, \link{mr_sens_z}, or \link{mr_sens_beta}}

\item{core_group}{vector containing groups that will be considered core instrument groups, default = 1}

\item{exclude_group}{vector containing groups that will be excluded when sampling instruments, default = 1}

\item{Nperm}{number of permutations}

\item{method_list}{MR methods, selected from TwoSampleMR::mr_method_list()}
}
\value{
data frame containing all MR estimates from permutations
}
\description{
Obtain distribution of MR estimates by randomly sampling instruments
}
