% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mr_sens.R
\name{plot_mr_sens}
\alias{plot_mr_sens}
\title{Plot results of sensitivity analysis}
\usage{
plot_mr_sens(mr_sens_result, scale = "linear")
}
\arguments{
\item{mr_sens_result}{output of \link{mr_sens_h2}, \link{mr_sens_z}, or \link{mr_sens_beta}}

\item{scale}{"linear": linear scale, or "exp": exponential scale}
}
\value{
ggplot object
}
\description{
Illustrate MR estimates by instrument group
}
