% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qstatistic.R
\name{Qstatistic}
\alias{Qstatistic}
\title{Q statistic}
\usage{
Qstatistic(b_group, se_group, b_overall)
}
\arguments{
\item{b_group}{a vector containing effect size estimates}

\item{se_group}{a vector containing standard errors of effect size estimates}

\item{b_overall}{overall effect size estimate}
}
\value{
list containing \itemize{
\item \code{Q} Cochran’s Q statistic
}
}
\description{
Cochran’s Q statistic
}
